Put type util in more central location
